etl/requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.6
elasticsearch==8.6.2
pydantic==1.10.7